+++ R/RunCARD.R
@@ -25,11 +25,7 @@
 #'
 #' @examples
 #' data(visium_human_pancreas_sub)
-#' spatial <- subset(
-#'   visium_human_pancreas_sub,
-#'   cells = colnames(visium_human_pancreas_sub)[1:120],
-#'   features = rownames(visium_human_pancreas_sub)[1:400]
-#' )
+#' spatial <- visium_human_pancreas_sub
 #' card_weights <- data.frame(
 #'   CARD_prop_Ductal = seq(0.70, 0.20, length.out = ncol(spatial)),
 #'   CARD_prop_Endocrine = seq(0.20, 0.70, length.out = ncol(spatial)),
@@ -88,7 +84,7 @@ RunCARD <- function(
   reference_layer = "counts",
   features = NULL,
   image = NULL,
-  coord.cols = c("col", "row"),
+  coord.cols = NULL,
   sample_varname = NULL,
   minCountGene = 100,
   minCountSpot = 5,