+++ R/RunCSIDE.R
@@ -41,11 +41,7 @@
 #'
 #' @examples
 #' data(visium_human_pancreas_sub)
-#' spatial <- subset(
-#'   visium_human_pancreas_sub,
-#'   cells = colnames(visium_human_pancreas_sub)[1:120],
-#'   features = rownames(visium_human_pancreas_sub)[1:400]
-#' )
+#' spatial <- visium_human_pancreas_sub
 #' spatial$region <- ifelse(spatial$x > stats::median(spatial$x), "right", "left")
 #' spatial$CSIDE_n_sig <- ifelse(spatial$region == "right", 12, 4)
 #' spatial$CSIDE_mode <- "regions"