+++ R/RunGiottoSpatialMethods.R
@@ -24,11 +24,7 @@
 #'
 #' @examples
 #' data(visium_human_pancreas_sub)
-#' spatial <- subset(
-#'   visium_human_pancreas_sub,
-#'   cells = colnames(visium_human_pancreas_sub)[1:120],
-#'   features = rownames(visium_human_pancreas_sub)[1:400]
-#' )
+#' spatial <- visium_human_pancreas_sub
 #' spatial$region <- ifelse(
 #'   spatial$x > stats::median(spatial$x),
 #'   "right",
@@ -71,7 +67,7 @@ RunGiottoCellProximity <- function(
   assay = NULL,
   layer = "data",
   image = NULL,
-  coord.cols = c("x", "y"),
+  coord.cols = NULL,
   network_method = c("Delaunay", "kNN"),
   network_name = NULL,
   number_of_simulations = 1000,
@@ -213,11 +209,7 @@ RunGiottoCellProximity <- function(
 #'
 #' @examples
 #' data(visium_human_pancreas_sub)
-#' spatial <- subset(
-#'   visium_human_pancreas_sub,
-#'   cells = colnames(visium_human_pancreas_sub)[1:120],
-#'   features = rownames(visium_human_pancreas_sub)[1:400]
-#' )
+#' spatial <- visium_human_pancreas_sub
 #' spatial <- Seurat::NormalizeData(spatial, assay = "Spatial", verbose = FALSE)
 #' giotto_genes <- list(
 #'   results = data.frame(
@@ -266,7 +258,7 @@ RunGiottoSpatialGenes <- function(
   layer = "data",
   features = NULL,
   image = NULL,
-  coord.cols = c("x", "y"),
+  coord.cols = NULL,
   network_method = c("Delaunay", "kNN"),
   network_name = NULL,
   bin_method = c("kmeans", "rank"),
@@ -418,11 +410,7 @@ RunGiottoSpatialGenes <- function(
 #'
 #' @examples
 #' data(visium_human_pancreas_sub)
-#' spatial <- subset(
-#'   visium_human_pancreas_sub,
-#'   cells = colnames(visium_human_pancreas_sub)[1:120],
-#'   features = rownames(visium_human_pancreas_sub)[1:400]
-#' )
+#' spatial <- visium_human_pancreas_sub
 #' module_features <- rownames(spatial)[1:4]
 #' module_cor <- expand.grid(
 #'   feat_ID = module_features,
@@ -473,7 +461,7 @@ RunGiottoSpatialModules <- function(
   layer = "data",
   features = NULL,
   image = NULL,
-  coord.cols = c("x", "y"),
+  coord.cols = NULL,
   network_method = c("Delaunay", "kNN"),
   network_name = NULL,
   cor_method = c("pearson", "spearman", "kendall"),