+++ R/RunSpaNorm.R
@@ -21,11 +21,7 @@
 #'
 #' @examples
 #' data(visium_human_pancreas_sub)
-#' spatial <- subset(
-#'   visium_human_pancreas_sub,
-#'   cells = colnames(visium_human_pancreas_sub)[1:80],
-#'   features = rownames(visium_human_pancreas_sub)[1:300]
-#' )
+#' spatial <- visium_human_pancreas_sub
 #' spatial <- Seurat::NormalizeData(spatial, assay = "Spatial", verbose = FALSE)
 #'
 #' SpatialSpotPlot(
@@ -67,7 +63,7 @@ RunSpaNorm <- function(
   assay = NULL,
   layer = "counts",
   image = NULL,
-  coord.cols = c("col", "row"),
+  coord.cols = NULL,
   new_assay = "SpaNorm",
   tool_name = "SpaNorm",
   store_results = TRUE,
@@ -168,7 +164,7 @@ spanorm_prepare_input <- function(
   assay,
   layer,
   image = NULL,
-  coord.cols = c("col", "row")
+  coord.cols = NULL
 ) {
   coords <- spatial_dim_coords(
     srt = srt,