+++ R/RunSpatialEcoTyper.R
@@ -84,11 +84,7 @@
 #'
 #' @examples
 #' data(visium_human_pancreas_sub)
-#' srt <- subset(
-#'   visium_human_pancreas_sub,
-#'   cells = colnames(visium_human_pancreas_sub)[1:120],
-#'   features = rownames(visium_human_pancreas_sub)[1:400]
-#' )
+#' srt <- visium_human_pancreas_sub
 #' srt$CellType <- srt$coda_label
 #' srt$SpatialEcoTyper_SE <- ifelse(srt$x > stats::median(srt$x), "SE1", "SE2")
 #' srt$sample <- ifelse(srt$y > stats::median(srt$y), "slice_a", "slice_b")
@@ -370,7 +366,7 @@ SpatialEcoTyperSpatialPlot <- function(
   image = NULL,
   overlay_image = TRUE,
   coord.cols = c(x.by, y.by),
-  palette = "Spectral",
+  palette = NULL,
   palcolor = NULL,
   theme_use = "theme_scop",
   ...